[refactor] Breakup io.cpp into smaller modules
