+++ src/cairo-utils.cpp
@@ -1,0 +1,158 @@
+#include "cairo-utils.hpp"
+
+#include <cairo/cairo.h>
+#include <inttypes.h>
+#include <math.h>
+#include <assert.h>
+
+#include <string>
+
+#include "attitude-utils.hpp"
+#include "decimal.hpp"
+#include "star-utils.hpp"
+
+namespace lost {
+
+/// Convert a colored Cairo image surface into a row-major array of grayscale pixels.
+/// Result is allocated with new[]
+unsigned char *SurfaceToGrayscaleImage(cairo_surface_t *cairoSurface) {
+    int width, height;
+    unsigned char *result;
+    uint32_t *cairoImage, pixel;
+
+    if (cairo_image_surface_get_format(cairoSurface) != CAIRO_FORMAT_ARGB32 &&
+        cairo_image_surface_get_format(cairoSurface) != CAIRO_FORMAT_RGB24) {
+        puts("Can't convert weird image formats to grayscale.");
+        return NULL;
+    }
+
+    width  = cairo_image_surface_get_width(cairoSurface);
+    height = cairo_image_surface_get_height(cairoSurface);
+
+    result = new unsigned char[width*height];
+    cairoImage = (uint32_t *)cairo_image_surface_get_data(cairoSurface);
+
+    for (int i = 0; i < height * width; i++) {
+        pixel = cairoImage[i];
+        // use "luminosity" method of grayscaling
+        result[i] = round(
+            (pixel>>16 &0xFF) *0.21 +
+            (pixel>>8  &0xFF) *0.71 +
+            (pixel     &0xFF) *0.07);
+    }
+
+    return result;
+}
+
+cairo_surface_t *GrayscaleImageToSurface(const unsigned char *image,
+                                         const int width, const int height) {
+    // cairo's 8-bit type isn't BW, it's an alpha channel only, which would look a bit weird lol
+    cairo_surface_t *result = cairo_image_surface_create(CAIRO_FORMAT_RGB24, width, height);
+    uint32_t *resultData = (uint32_t *)cairo_image_surface_get_data(result);
+    // hopefully unnecessary
+    cairo_surface_flush(result);
+    for (long i = 0; i < width * height; i++) {
+        // equal r, g, and b components
+        resultData[i] = (image[i] << 16) + (image[i] << 8) + (image[i]);
+    }
+    cairo_surface_mark_dirty(result);
+    return result;
+}
+
+/**
+ * Plot information about an image onto the image using Cairo.
+ * Puts a box around each centroid, writes the attitude in the top left, etc.
+ * @param red,green,blue The color to use when annotating the image. 0 represents none of that color, 1 represents that color in full.
+ * @param alpha The transparency of annotations. 0 is completely transparent, 1 is completely opaque.
+ * @param rawStarIndexes If true, print the catalog index. This is in contrast with the default behavior, which is to print the name of the catalog star.
+ */
+void SurfacePlot(std::string description,
+                 cairo_surface_t *cairoSurface,
+                 const Stars &stars,
+                 const StarIdentifiers *starIds,
+                 const Catalog *catalog,
+                 const Attitude *attitude,
+                 double red,
+                 double green,
+                 double blue,
+                 double alpha,
+                 // if true, don't use catalog name
+                 bool rawStarIndexes) {
+    cairo_t *cairoCtx;
+    std::string metadata = description + " ";
+
+    cairoCtx = cairo_create(cairoSurface);
+    cairo_set_source_rgba(cairoCtx, red, green, blue, alpha);
+    cairo_set_line_width(cairoCtx, 1.0);
+    cairo_set_antialias(cairoCtx, CAIRO_ANTIALIAS_NONE);
+    cairo_font_options_t *cairoFontOptions = cairo_font_options_create();
+    cairo_font_options_set_antialias(cairoFontOptions, CAIRO_ANTIALIAS_NONE);
+    cairo_set_font_options(cairoCtx, cairoFontOptions);
+    cairo_text_extents_t cairoTextExtents;
+    cairo_text_extents(cairoCtx, "1234567890", &cairoTextExtents);
+    decimal textHeight = cairoTextExtents.height;
+
+    for (const Star &centroid : stars) {
+        // plot the box around the star
+        if (centroid.radiusX > DECIMAL(0.0)) {
+            decimal radiusX = centroid.radiusX;
+            decimal radiusY = centroid.radiusY > DECIMAL(0.0) ?
+                centroid.radiusY : radiusX;
+
+            // Rectangles should be entirely /outside/ the radius of the star, so the star is
+            // fully visible.
+            cairo_rectangle(cairoCtx,
+                            centroid.position.x - radiusX,
+                            centroid.position.y - radiusY,
+                            radiusX * 2,
+                            radiusY * 2);
+            cairo_stroke(cairoCtx);
+        } else {
+            cairo_rectangle(cairoCtx,
+                            DECIMAL_FLOOR(centroid.position.x),
+                            DECIMAL_FLOOR(centroid.position.y),
+                            1, 1);
+            cairo_fill(cairoCtx);
+        }
+    }
+
+    metadata += std::to_string(stars.size()) + " centroids   ";
+
+    if (starIds != NULL) {
+        assert(catalog != NULL);
+
+        for (const StarIdentifier &starId : *starIds) {
+            const Star &centroid = stars[starId.starIndex];
+            cairo_move_to(cairoCtx,
+
+                          centroid.radiusX > DECIMAL(0.0)
+                          ? centroid.position.x + centroid.radiusX + 3
+                          : centroid.position.x + 8,
+
+                          centroid.radiusY > DECIMAL(0.0)
+                          ? centroid.position.y - centroid.radiusY + textHeight
+                          : centroid.position.y + 10);
+
+            int plotName = rawStarIndexes ? starId.catalogIndex : (*catalog)[starId.catalogIndex].name;
+            cairo_show_text(cairoCtx, std::to_string(plotName).c_str());
+        }
+        metadata += std::to_string(starIds->size()) + " identified   ";
+    }
+
+    if (attitude != NULL) {
+        EulerAngles spherical = attitude->ToSpherical();
+        metadata +=
+            "RA: " + std::to_string(RadToDeg(spherical.ra)) + "  " +
+            "DE: " + std::to_string(RadToDeg(spherical.de)) + "  " +
+            "Roll: " + std::to_string(RadToDeg(spherical.roll)) + "   ";
+    }
+
+    // plot metadata
+    cairo_move_to(cairoCtx, 3, 3 + textHeight);
+    cairo_show_text(cairoCtx, metadata.c_str());
+
+    cairo_font_options_destroy(cairoFontOptions);
+    cairo_destroy(cairoCtx);
+}
+
+} // namespace lost
