+++ src/pipeline-input.hpp
@@ -1,0 +1,192 @@
+#ifndef PIPELINE_INPUT_H
+#define PIPELINE_INPUT_H
+
+#include <cairo/cairo.h>
+
+#include <random>
+#include <vector>
+#include <string>
+#include <memory>
+
+#include "centroiders.hpp"
+#include "star-utils.hpp"
+#include "star-id.hpp"
+#include "camera.hpp"
+#include "attitude-utils.hpp"
+#include "attitude-estimators.hpp"
+#include "databases.hpp"
+#include "io-util.hpp"
+#include "cairo-utils.hpp"
+
+#ifndef CAIRO_HAS_PNG_FUNCTIONS
+#error LOST requires Cairo to be compiled with PNG support
+#endif
+
+namespace lost {
+
+////////////////////
+// PIPELINE INPUT //
+////////////////////
+
+/// The command line options passed when running a pipeline
+class PipelineOptions {
+public:
+#define LOST_CLI_OPTION(name, type, prop, defaultVal, converter, defaultArg) \
+    type prop = defaultVal;
+#include "./pipeline-options.hpp"
+#undef LOST_CLI_OPTION
+};
+
+/**
+ * Represents the input and expected outputs of a pipeline run.
+ * This is all the data about the pipeline we are about to run that can be gathered without actually running the pipeline. The "input" members return the things that will be fed into the pipeline. The "expected" methods return the "correct" outputs, i.e. what a perfect star tracking algorithm would output (this is only meaningful whe the image is generated, otherwise we don't know the correct output!). The "expected" methods are used to evaluate the quality of our algorithms. Some of the methods (both input and expected) may return NULL for certain subclasses.
+ * By default, the "expected" methods return the corresponding inputs, which is reasonable behavior unless you are trying to intentionally introduce error into the inputs.
+ */
+class PipelineInput {
+public:
+    virtual ~PipelineInput(){};
+
+    virtual const Image *InputImage() const { return NULL; };
+    /// The catalog to which catalog indexes returned from other methods refer.
+    virtual const Catalog &GetCatalog() const = 0;
+    virtual const Stars *InputStars() const { return NULL; };
+    /// The centroid indices in the StarIdentifiers returned from InputStarIds should be indices
+    /// into InputStars(), not ExpectedStars(), when present, because otherwise it's useless.
+    virtual const StarIdentifiers *InputStarIds() const { return NULL; };
+    /// Only used in tracking mode, in which case it is an estimate of the current attitude based on the last attitude, IMU info, etc.
+    virtual const Attitude *InputAttitude() const { return NULL; };
+    virtual const Camera *InputCamera() const { return NULL; };
+    /// Convert the InputImage() output into a cairo surface
+    cairo_surface_t *InputImageSurface() const;
+
+    virtual const Stars *ExpectedStars() const { return InputStars(); };
+    /// Centroid indices in the StarIdentifiers returned from ExpectedStarIds should be indices into
+    /// ExpectedStars(), /not/ InputStars(). This is in contrast to InputStarIds. If you need to
+    /// compare ExpectedStarIds against the input stars, then you should use some function which
+    /// uses simple heuristics to match the input stars and expected stars (eg based on distance).
+    /// Cf how the star-ID comparator works for a reference implementation.
+    virtual const StarIdentifiers *ExpectedStarIds() const { return InputStarIds(); };
+    virtual const Attitude *ExpectedAttitude() const { return InputAttitude(); };
+};
+
+/**
+ * A pipeline input which is generated (fake image).
+ * Uses a pretty decent image generation algorithm to create a fake image as the basis for a pipeline input. Since we know everything about the generated image, all of the input and expected methods are available.
+ */
+class GeneratedPipelineInput : public PipelineInput {
+public:
+    GeneratedPipelineInput(const Catalog &, Attitude, Camera, std::default_random_engine *,
+
+                           bool centroidsOnly,
+                           decimal observedReferenceBrightness, decimal starSpreadStdDev,
+                           decimal sensitivity, decimal darkCurrent, decimal readNoiseStdDev,
+                           Attitude motionBlurDirection, decimal exposureTime, decimal readoutTime,
+                           bool shotNoise, int oversampling,
+                           int numFalseStars, int falseMinMagnitude, int falseMaxMagnitude,
+                           int cutoffMag,
+                           decimal perturbationStddev);
+
+
+    const Image *InputImage() const override { return &image; };
+    const Stars *InputStars() const override { return &inputStars; };
+    const Stars *ExpectedStars() const override { return &expectedStars; };
+    const Camera *InputCamera() const override { return &camera; };
+    const StarIdentifiers *InputStarIds() const override { return &inputStarIds; };
+    const StarIdentifiers *ExpectedStarIds() const override { return &expectedStarIds; };
+    const Attitude *InputAttitude() const override { return &attitude; };
+    const Catalog &GetCatalog() const override { return catalog; };
+
+private:
+    std::vector<unsigned char> imageData;
+    Image image;
+    /// Includes false stars and very dim stars. Any further filtering that needs to happen before comparison happens in the comparator itself.
+    Stars expectedStars;
+    /// Includes perturbations, filtered down to magnitude, etc. Whatever the star-id algorithm needs.
+    Stars inputStars;
+    Camera camera;
+    Attitude attitude;
+    const Catalog &catalog;
+    StarIdentifiers inputStarIds;
+    StarIdentifiers expectedStarIds;
+};
+
+typedef std::vector<std::unique_ptr<PipelineInput>> PipelineInputList;
+
+PipelineInputList GetPipelineInput(const PipelineOptions &values);
+
+/// A pipeline input created by reading a PNG from a file on disk.
+class PngPipelineInput : public PipelineInput {
+public:
+    PngPipelineInput(cairo_surface_t *, Camera, const Catalog &);
+    ~PngPipelineInput();
+
+    const Image *InputImage() const override { return &image; };
+    const Camera *InputCamera() const override { return &camera; };
+    const Catalog &GetCatalog() const override { return catalog; };
+
+private:
+    Image image;
+    Camera camera;
+    const Catalog &catalog;
+};
+
+/////////////////////
+// PIPELINE OUTPUT //
+/////////////////////
+
+/**
+ * @brief The result of running a pipeline.
+ * @details Also stores intermediate outputs, not just the final attitude.
+ */
+struct PipelineOutput {
+    std::unique_ptr<Stars> stars = nullptr;
+    std::unique_ptr<StarIdentifiers> starIds = nullptr;
+    std::unique_ptr<Attitude> attitude = nullptr;
+
+    /// How many nanoseconds the centroiding stage of the pipeline took. Similarly for the other
+    /// fields. If negative, the centroiding stage was not run.
+    long long centroidingTimeNs = -1;
+    long long starIdTimeNs = -1;
+    long long attitudeEstimationTimeNs = -1;
+
+    /**
+     * @brief The catalog that the indices in starIds refer to
+     * @todo Don't store it here
+     */
+    Catalog catalog;
+};
+
+//////////////
+// PIPELINE //
+//////////////
+
+/**
+ * @brief A set of algorithms that describes all or part of the star-tracking "pipeline"
+ * @details A centroiding algorithm identifies the (x,y) pixel coordinates of each star detected in the raw image. The star id algorithm then determines which centroid corresponds to which catalog star. Finally, the attitude estimation algorithm determines the orientation of the camera based on the centroids and identified stars.
+ */
+class Pipeline {
+    friend Pipeline SetPipeline(const PipelineOptions &values);
+
+public:
+    Pipeline() = default;
+    Pipeline(CentroidAlgorithm *, StarIdAlgorithm *, AttitudeEstimationAlgorithm *, unsigned char *);
+    PipelineOutput Go(const PipelineInput &);
+    std::vector<PipelineOutput> Go(const PipelineInputList &);
+
+private:
+    std::unique_ptr<CentroidAlgorithm> centroidAlgorithm;
+
+    // next two options are for magnitude filter:
+    int centroidMinMagnitude = 0;
+    int centroidMinStars = 0;
+
+    std::unique_ptr<StarIdAlgorithm> starIdAlgorithm;
+    std::unique_ptr<AttitudeEstimationAlgorithm> attitudeEstimationAlgorithm;
+    std::unique_ptr<unsigned char[]> database;
+};
+
+Pipeline SetPipeline(const PipelineOptions &values);
+
+}
+
+#endif
